# fetch-api/connectors/notifications/docs/swagger.yaml
basePath: /
definitions:
  notifications-controller_internal_http_dto_request.DownloaderNotificationPayload:
    properties:
      actions:
        items:
          $ref: '#/definitions/notifications-controller_internal_http_dto_request.NotificationAction'
        type: array
      category:
        type: string
      channel:
        type: string
      event:
        type: string
      fields:
        items:
          $ref: '#/definitions/notifications-controller_internal_http_dto_request.NotificationField'
        type: array
      hash:
        type: string
      message:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      severity:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  notifications-controller_internal_http_dto_request.GrafanaAlert:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      dashboardURL:
        type: string
      endsAt:
        type: string
      fingerprint:
        type: string
      generatorURL:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      panelURL:
        type: string
      silenceURL:
        type: string
      startsAt:
        type: string
      status:
        type: string
      valueString:
        type: string
    type: object
  notifications-controller_internal_http_dto_request.GrafanaNotificationPayload:
    properties:
      alerts:
        items:
          $ref: '#/definitions/notifications-controller_internal_http_dto_request.GrafanaAlert'
        type: array
      channel:
        type: string
      commonAnnotations:
        additionalProperties:
          type: string
        type: object
      commonLabels:
        additionalProperties:
          type: string
        type: object
      externalURL:
        type: string
      groupKey:
        type: string
      groupLabels:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      orgId:
        type: integer
      receiver:
        type: string
      state:
        type: string
      status:
        type: string
      title:
        type: string
      truncatedAlerts:
        type: integer
      version:
        type: string
    type: object
  notifications-controller_internal_http_dto_request.NotificationAction:
    properties:
      id:
        type: string
      label:
        type: string
      style:
        type: string
      url:
        type: string
      value:
        type: string
    type: object
  notifications-controller_internal_http_dto_request.NotificationField:
    properties:
      short:
        type: boolean
      title:
        type: string
      value:
        type: string
    type: object
  notifications-controller_internal_http_dto_response.ErrorResponse:
    properties:
      error:
        type: string
      upstream_response:
        additionalProperties: {}
        type: object
    type: object
  notifications-controller_internal_http_dto_response.HealthResponse:
    properties:
      connector_name:
        type: string
      health_last_check:
        type: string
      health_next_check:
        type: string
      healthy:
        type: boolean
    type: object
  notifications-controller_internal_http_dto_response.NotificationDeliveryResponse:
    properties:
      channel:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  notifications-controller_internal_http_dto_response.ReadyResponse:
    properties:
      ready:
        type: boolean
    type: object
info:
  contact: {}
  description: A connector that acts as a notifications controller that accepts payloads
    and sends notifications to different channels.
  title: Notifications Connector
paths:
  /api/health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.HealthResponse'
      summary: Health check
      tags:
      - health
  /api/ready:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.ReadyResponse'
      summary: Readiness check
      tags:
      - health
  /notifications/downloader:
    post:
      consumes:
      - application/json
      description: Accepts a downloader-originated notification payload and posts
        it to Slack.
      parameters:
      - description: Downloader notification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications-controller_internal_http_dto_request.DownloaderNotificationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.NotificationDeliveryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.ErrorResponse'
      summary: Forward a connector-downloader event payload to Slack
      tags:
      - notifications
  /notifications/grafana:
    post:
      consumes:
      - application/json
      description: Accepts a Grafana Alerting webhook payload and posts it to Slack.
      parameters:
      - description: Grafana notification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications-controller_internal_http_dto_request.GrafanaNotificationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.NotificationDeliveryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/notifications-controller_internal_http_dto_response.ErrorResponse'
      summary: Forward a Grafana alert payload to Slack
      tags:
      - notifications
produces:
- application/json
swagger: "2.0"